requirements.txt
# torch==2.1.2

numpy>=1.26.0
rouge_score>=0.1.2
fire>=0.5.0

transformers==4.45.2
sentencepiece>=0.1.99
tokenizers==0.20.1

accelerate>=0.34.0
datasets>=2.14.5
deepspeed==0.14.0
peft>=0.5.0

einops>=0.7.0
bitsandbytes==0.41.1
scipy>=1.11.3
protobuf>=4.24.4
torchmetrics>=1.2.0


